3: Extract `Wrapped/ElementList` from `Gallery`
